+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/modules/oppgaver/endret-sluttdato/i18n/nb.ts
@@ -11,17 +11,4 @@ export const endretSluttdatoMessages_nb = {
         'Hvis vi ikke hører fra deg innen svarfristen har gått ut, bruker vi {formatertDato} som sluttdato når vi behandler saken din.',
     'endretSluttdato.oppsummering.endret':
         'Din sluttdato i ungdomsprogrammet er endret fra <strong>{forrige}</strong> til <strong>{ny}</strong>.',
-    'endretSluttdato.oppsummering.meldtUt': 'Din sluttdato i ungdomsprogrammet er satt til <strong>{ny}</strong>.',
-
-    'endretSluttdato.meldUtOppgavetekst.1':
-        'Veilederen din har meldt deg ut av ungdomsprogrammet med sluttdato <strong>{formatertDato}</strong>.',
-    'endretSluttdato.meldUtOppgavetekst.2':
-        'Du får denne meldingen slik at du kan komme med en tilbakemelding på datoen.',
-    'endretSluttdato.meldUtOppgavetekst.3':
-        'Ingen tilbakemelding? Kryss av på “Nei” med en gang og send inn svaret ditt. Jo fortere du svarer, jo fortere får vi behandlet saken din.',
-    'endretSluttdato.meldUtOppgavetekst.4':
-        'Har du en tilbakemelding? Ta kontakt med veilederen din først. Når dere har snakket sammen, sender du inn svaret ditt her.',
-    'endretSluttdato.meldUtOppgavetekst.svarfrist': 'Fristen for å svare er senest {formatertFrist}.',
-    'endretSluttdato.meldUtOppgavetekst.5':
-        'Hvis vi ikke hører fra deg innen svarfristen har gått ut, bruker vi {formatertDato} som sluttdato når vi behandler saken din.',
 };