+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @navikt/sif-common-ui
 
+## 0.0.96
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/sif-common-formik-ds@2.5.43
+    - @navikt/sif-common-core-ds@10.3.51
+    - @navikt/sif-common-utils@3.49.59
+    - @navikt/sif-validation@1.0.18
+
 ## 0.0.95
 
 ### Patch Changes