+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-forms-ds
 
+## 21.0.53
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/sif-common-formik-ds@2.5.43
+    - @navikt/sif-common-core-ds@10.3.51
+    - @navikt/sif-common-utils@3.49.59
+    - @navikt/sif-common-ui@0.11.52
+    - @navikt/sif-validation@1.0.18
+
 ## 21.0.52
 
 ### Patch Changes