+++ packages/sif-common-soknad-ds/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @navikt/sif-common-soknad-ds
 
+## 26.1.35
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/sif-common-formik-ds@2.5.43
+    - @navikt/sif-common-forms-ds@21.0.53
+    - @navikt/appstatus-react-ds@2.20.70
+    - @navikt/sif-common-core-ds@10.3.51
+    - @navikt/sif-common-sentry@0.28.61
+    - @navikt/sif-common-utils@3.49.59
+    - @navikt/sif-common-faro@0.2.11
+    - @navikt/sif-common-ui@0.11.52
+    - @navikt/sif-common-hooks@0.3.8
+    - @navikt/sif-validation@1.0.18
+
 ## 26.1.34
 
 ### Patch Changes